+++ static/js/index.js
@@ -4,7 +4,7 @@ import "regenerator-runtime/runtime";
 // jQuery and selectWoo are very special beasts
 // https://stackoverflow.com/a/47984928
 // https://stackoverflow.com/a/49722514
-import jquery from "jquery";
+import jquery, { Autocomplete } from "jquery";
 window.$ = window.jQuery = jquery;
 import select2 from "../vendor/selectWoo-1.0.8/js/select2.full.min";
 select2(window.$);
@@ -20,4 +20,25 @@ import "leaflet-center-cross";
 import "chart.js";
 import "sentry";
 
+// app modules
+import forms from "./forms";
+import map from "./map";
+import search from "./search";
+
 import "./app";
+
+window.a4a = {
+  forms,
+  map,
+  search,
+};
+
+window.addEventListener("DOMContentLoaded", () => {
+  // Django crispy forms asterisk a11y improvements
+  // XXX: move me to another module
+  $(".asteriskField").each((i, elem) => {
+    $(elem).replaceWith('&nbsp;<small>(requis)</small><abbr class="asteriskField" title="(obligatoire)">*</abbr>');
+  });
+
+  map.handleGeoLinks();
+});
